Update CommandsParser.py
I'm not sure its recommended or possible to try to grab an attribute of a class that doesn't have an instantiation. If we would like to keep going down this road, we'd need to change the type json attribute to reflect the class name. Could to_lower() the type so it doesn't have to be camel cased, but i'm out of ideas on how to make it more elegant than that.